Extract Prompt Studio value sync helpers
