+++ tests/test_frontend_modules.py
@@ -82,6 +82,7 @@ class FrontendModuleStructureTests(unittest.TestCase):
         self.assertIn('./prompt_studio/layout.js"', source)
         self.assertIn('./prompt_studio/advanced_layout_controller.js"', source)
         self.assertIn('./prompt_studio/studio_textareas.js"', source)
+        self.assertIn('./prompt_studio/studio_values.js"', source)
         self.assertIn('./textarea.js"', advanced_fields_ui_source)
         self.assertIn('./prompt_studio/wheel.js"', source)
         self.assertIn('./prompt_studio/serialization.js"', source)
@@ -152,6 +153,9 @@ class FrontendModuleStructureTests(unittest.TestCase):
         ).read_text(encoding="utf-8")
         studio_textareas_source = (
             PROMPT_STUDIO_MODULES / "studio_textareas.js"
+        ).read_text(encoding="utf-8")
+        studio_values_source = (
+            PROMPT_STUDIO_MODULES / "studio_values.js"
         ).read_text(encoding="utf-8")
         textarea_source = (PROMPT_STUDIO_MODULES / "textarea.js").read_text(
             encoding="utf-8"
@@ -392,6 +396,15 @@ class FrontendModuleStructureTests(unittest.TestCase):
                 self.assertIn(f"  {name},", studio_textareas_source)
 
         for name in (
+            "applyExecutedInputs",
+            "restoreInputFromWidget",
+            "syncStudioValues",
+            "syncWidgetValue",
+        ):
+            with self.subTest(module="studio_values", symbol=name):
+                self.assertIn(f"  {name},", studio_values_source)
+
+        for name in (
             "advancedFieldTextareaPlaceholder",
             "advancedFieldTextareaTitle",
             "captureAdvancedTextareaManualResize",
@@ -444,6 +457,7 @@ class FrontendModuleStructureTests(unittest.TestCase):
             "widgets.js",
             "layout.js",
             "studio_textareas.js",
+            "studio_values.js",
             "textarea.js",
             "wheel.js",
             "serialization.js",
